// cgocore-ui/src/app/features/models/loading/filters/DatedFlightUldFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightUldFilter {

datedFlightCompanyIataCode: string;
datedFlightFlightNumber: string;
datedFlightDepartureDateMinValue: string;
datedFlightDepartureDateMaxValue: string;
uldTypeCode: string;
uldCode: string;
uldOwnerCode: string;

/* Specific Code Start */
/* Specific Code End */
}